Update patch set 4
Patch Set 4:

I've used two things to look at this:
1) https://daringfireball.net/projects/markdown/dingus
2) Markdown preview in vscode

* some bullet point
on a new line

if you leave out the two spaces, it should (according to specification) end up on one line:
If you put two spaces at the end of the line, it will end up rendered in two lines.

I'm not sure I understand the problem here. I've referenced the official spec, and this is a language element.  If there are mistakes in how I've used this language element, I'll gladly fix those.

dillinger.io does not render markdown properly, specifically the concept of block continuation.  Try this: 

this line
should not break

it will render on two lines, and should not.

Patch-set: 4