Provide for UCP standards and conventions
This change intends to account for various standards and conventions
used by the undercloud components. As the topic of conventions for
any active project is naturally open ended, this set of changes does
not intend to be a complete representation.

Several of the sections identified for conventions are marked as under
development, and will not be formalized as part of this change.

Change-Id: Ideaa9ed4146f4a6dcd2d3c44037bcac2ed3755bb